rename : 전체 폴더 변경 com.example.givemeticon.** -> com.givemeticon.**
