--- src/main/java/com/example/givemeticon/GivemeticonApplication.java
@@ -1,13 +1,0 @@
-package com.example.givemeticon;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class GivemeticonApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(GivemeticonApplication.class, args);
-	}
-
-}
