--- src/main/java/com/example/givemeticon/common/response/ErrorResult.java
@@ -1,8 +1,0 @@
-package com.example.givemeticon.common.response;
-
-/**
- * 에러 응답 형식
- */
-public record ErrorResult(String message) {
-
-}
