--- src/main/java/com/example/givemeticon/user/domain/UserRole.java
@@ -1,5 +1,0 @@
-package com.example.givemeticon.user.domain;
-
-public enum UserRole {
-    USER, VIP
-}
