--- src/main/java/com/example/givemeticon/user/presentation/UserController.java
@@ -1,28 +1,0 @@
-package com.example.givemeticon.user.presentation;
-
-import com.example.givemeticon.user.application.UserService;
-import com.example.givemeticon.user.presentation.request.SignUpRequest;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.validation.annotation.Validated;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/api/v1/users")
-public class UserController {
-
-    private final UserService userService;
-
-    public UserController(UserService userService) {
-        this.userService = userService;
-    }
-
-    @PostMapping("/sign-up")
-    public ResponseEntity<Void> signUp(@RequestBody @Validated SignUpRequest request) {
-        userService.signUp(request);
-        return new ResponseEntity<>(HttpStatus.CREATED);
-    }
-}
