+++ src/main/java/com/jinddung2/givemeticon/GivemeticonApplication.java
@@ -1,0 +1,13 @@
+package com.jinddung2.givemeticon;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class GivemeticonApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(GivemeticonApplication.class, args);
+    }
+
+}
