+++ src/main/java/com/jinddung2/givemeticon/common/response/ErrorResult.java
@@ -1,0 +1,8 @@
+package com.jinddung2.givemeticon.common.response;
+
+/**
+ * 에러 응답 형식
+ */
+public record ErrorResult(String message) {
+
+}
