+++ src/main/java/com/jinddung2/givemeticon/user/domain/UserRole.java
@@ -1,0 +1,5 @@
+package com.jinddung2.givemeticon.user.domain;
+
+public enum UserRole {
+    USER, VIP
+}
