+++ src/main/java/com/jinddung2/givemeticon/user/exception/DuplicatedEmailException.java
@@ -1,0 +1,10 @@
+package com.jinddung2.givemeticon.user.exception;
+
+import com.jinddung2.givemeticon.common.exception.ErrorCode;
+
+public class DuplicatedEmailException extends UserException {
+
+    public DuplicatedEmailException() {
+        super(ErrorCode.DUPLICATED_EMAIL);
+    }
+}
