+++ src/main/resources/mapper/UserMapper.xml
@@ -2,7 +2,7 @@
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="com.example.givemeticon.user.infrastructure.mapper.UserMapper">
+<mapper namespace="com.jinddung2.givemeticon.user.infrastructure.mapper.UserMapper">
     <insert id="save" parameterType="User" useGeneratedKeys="true" keyProperty="id">
         INSERT INTO user (id,
                           account_id,
