--- src/test/java/com/example/givemeticon/GivemeticonApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.givemeticon;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class GivemeticonApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
