--- src/test/java/com/example/givemeticon/resources/application.yml
@@ -1,11 +1,0 @@
-spring:
-  datasource:
-    url: jdbc:mysql://localhost:3307/givemeticon-db
-    username: root
-    password: 1234
-
-mybatis:
-  mapper-locations: classpath:mapper/**/*.xml
-  type-aliases-package: com.example.givemeticon.*.domain
-  configuration:
-    map-underscore-to-camel-case: true
