+++ src/test/java/com/jinddung2/givemeticon/GivemeticonApplicationTests.java
@@ -1,0 +1,11 @@
+package com.jinddung2.givemeticon;
+
+import org.junit.jupiter.api.Test;
+
+class GivemeticonApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
